refactor!(core): Move include files into gprat/ subdirectory
This is a separate commit for git's rename tracking
