+++ core/include/gprat/cpu/adapter_cblas_fp32.hpp
