+++ core/include/gprat/cpu/gp_algorithms.hpp
