+++ core/include/gprat/cpu/gp_functions.hpp
