+++ core/include/gprat/cpu/gp_uncertainty.hpp
