+++ core/include/gprat/gp_kernels.hpp
