+++ core/include/gprat/utils_c.hpp
