+++ test/src/output_correctness.cpp
@@ -1,5 +1,5 @@
-#include "gprat_c.hpp"
-#include "utils_c.hpp"
+#include "gprat/gprat_c.hpp"
+#include "gprat/utils_c.hpp"
 #include <catch2/catch_test_macros.hpp>
 #include <catch2/matchers/catch_matchers_floating_point.hpp>
